Extracted components/Request.js components into separate files
